+++ react/package.json
@@ -8,11 +8,11 @@
     "@testing-library/react": "^9.3.2",
     "@testing-library/user-event": "^7.1.2",
     "bootstrap": "^4.5.0",
-    "react-bootstrap": "^1.0.0-beta.12",
     "react": "^16.13.1",
+    "react-bootstrap": "^1.0.0-beta.12",
     "react-dom": "^16.13.1",
     "react-router-dom": "^5.2.0",
-    "react-scripts": "3.4.1"
+    "react-scripts": "^3.4.3"
   },
   "scripts": {
     "start": "react-scripts start",
