+++ react/src/index.css
@@ -1,13 +1,17 @@
+/* Global styles */
+
+/* Colos scheme from: https://coolors.co/000000-14213d-fca311-e5e5e5-ffffff*/
+
 :root{
-  --color-black: #011627;
-  --color-white: #fdfffc;
-  --color-blue: #2ec4b6;
-  --color-red: #e71d36;
-  --color-orange: #ff9f1c;
+  --color-black: #000000;
+  --color-white: #FFFFFF;
+  --color-gray: #E5E5E5;
+  --color-blue: #14213D;
+  --color-orange: #FCA311;
 }
 
 html, body, #app, #app>div {
-  height: 100%
+  height: 100vh;
 }
 
 body{
