--- react/src/pages/TecMM/Navbartec.css
@@ -1,9 +1,0 @@
-.link{
-    color: var(--color-black); /* color from index.css*/
-    font-size: 18px;
-}
-
-.link:hover{
-    color: var(--color-blue); /* color from index.css*/
-    transition: all .3s;
-}
