--- react/src/pages/TecMM/Tecmm.css
@@ -1,113 +1,0 @@
-/* --------------------------------------------------------------------------- Tecmm page general styles*/
-.tecmm-page {
-  background-color: #faf8ff;
-  color: #474554;
-}
-
-.section {
-  margin-top: 50px;
-}
-
-.title-section {
-  margin-bottom: 30px;
-  font-size: 35px;
-}
-
-/* --------------------------------------------------------------------------- Navbar styles */
-
-.navbar {
-  background-color: #faf8ff;
-  color: #474554;
-}
-
-.link {
-  color: #474554;
-  font-size: 18px;
-}
-
-.link:hover {
-  color: black;
-  transition: all 0.3s;
-}
-
-/* --------------------------------------------------------------------------- Hero section*/
-
-.section-hero {
-  background-image: url("../../img/tec-chapala.jpg");
-  background-position: center;
-  background-repeat: no-repeat;
-  background-size: cover;
-  min-height: 95vh;
-  color: #faf8ff;
-  margin-top: 56px;
-}
-
-/* --------------------------------------------------------------------------- Events section*/
-
-.masonry {
-  display: grid;
-  grid-template-columns: repeat(auto-fill, 250px);
-  grid-gap: 5px;
-  justify-content: center;
-}
-
-.item {
-  border-radius: 5px;
-  padding: 10px;
-}
-
-.item img {
-  max-width: 100%;
-}
-
-.level-1 {
-  grid-row-end: span 3;
-}
-
-.level-2 {
-  grid-row-end: span 2;
-}
-
-.level-3 {
-  grid-row-end: span 1;
-}
-
-/* --------------------------------------------------------------------------- Anouncement styles*/
-
-.important {
-  color: #ff0000;
-}
-
-.section-links-images {
-  background-color: #fff0ea;
-  padding-right: 0px !important;
-  padding-left: 0px !important;
-}
-
-.section-links-images img {
-  width: 100%;
-  height: 100%;
-}
-
-.section-links-text {
-  padding: 25px;
-  font-weight: bold;
-}
-
-.link-title {
-  font-size: 25px;
-}
-
-.link-color-1 {
-  background-color: #ffe8f7;
-}
-
-.link-color-2 {
-  background-color: #ffdfca;
-}
-
-.link-color-3 {
-  background-color: #ffe694;
-}
-
-/* --------------------------------------------------------------------------- Footer styles */
